refactor(dvfy-tree-node): extract into dedicated file
DvfyTreeNode was co-located with DvfyTreeView in dvfy-tree-view.js.
Split into components/dvfy-tree-node.js to match the one-component-per-file
convention used by every other component.

No public API change: same custom element tag, attributes, events, slots.
Node-specific CSS moves with the class; tree-view.js keeps container +
focus styles only. Regenerated custom-elements.json reflects new path.

Refs #318
